+++ src/map/sv_map.jl
@@ -23,8 +23,8 @@ function Params_SV(img; periodic::Bool=true)
     p = Params_SV(
         periodic,
         total,
-        similar(img, ComplexF64, complex_box),
-        similar(img, ComplexF64, complex_box)
+        similar(img, ComplexF32, complex_box),
+        similar(img, ComplexF32, complex_box)
     )
     cf_type = :full
     p, cf_type
@@ -63,7 +63,7 @@ The `image` contains the probability of the voxel being in the correct phase.
 # Examples
 ```jldoctest
 julia> surfvoid([1 0; 0 1]; periodic=true).result
-3×3 Matrix{Float64}:
+3×3 Matrix{Float32}:
  0.176777  0.176777  0.176777
  0.176777  0.176777  0.176777
  0.176777  0.176777  0.176777
